+++ index.js
@@ -1,19 +1,19 @@
 /**
- * Full fhirconvert function: validates, converts, then identifies variables.
+ * Full fhirConvert function: validates, converts, then identifies variables.
  * @param {string} str - inputted normal syntax expression
  * @param {Array} vars - array of usable variables entered by user
  * @returns converted fhirpath expression
  */
- export function fhirconvert(str, vars) {
+export function fhirConvert(str, vars) {
   if (validate(str, vars)) {
-    return varfind(convert(str), vars);
+    return varFind(convert(str), vars);
   } else {
     return null;
   }
 }
 
 // Array of usable operators
-let ops = [
+const ops = [
   "+",
   "-",
   "*",
@@ -40,7 +40,7 @@ let ops = [
 ];
 
 // Array of usable functions
-let funs = [
+const functions = [
   "CEILING",
   "FLOOR",
   "ABS",
@@ -64,7 +64,7 @@ let funs = [
 ];
 
 // Array of functions with no arguments
-let funs2 = [
+const noArgumentFunctions = [
   "CEILING",
   "FLOOR",
   "ABS",
@@ -92,16 +92,16 @@ let funs2 = [
  * @param {Array} vars - array of usable variables
  * @returns boolean, valid or invalid
  */
-export function validate(str, vars) {
+function validate(str, vars) {
   // Operator Validation
-  var len = str.length;
-  var op = "";
+  let len = str.length;
+  let op = "";
   // Loop to identify operator strings
-  for (var j = 0; j < len; j++) {
+  for (let j = 0; j < len; j++) {
     // If operator char, append to op
     if (!(/[a-zA-Z0-9.,'"\s()\\-]/.test(str[j]))) {
       op = op + str[j];
-      if (j == len - 1 || j == 0) {
+      if (j === len - 1 || j === 0) {
         return false;
       }
     }
@@ -111,30 +111,30 @@ export function validate(str, vars) {
         return false;
       }
       // Validates expression to left of operator
-      var lsearch = true;
-      var op_l = j - op.length - 1;
-      while(lsearch) {
+      let lSearch = true;
+      let op_l = j - op.length - 1;
+      while(lSearch) {
         if (!(/[\s]/.test(str[op_l]))) {
           if (!(/[a-zA-Z0-9.,'")\\-]/.test(str[op_l]))) {
             return false;
           }
-          lsearch = false;
+          lSearch = false;
         } else {
           op_l -= 1;
         }
       }
       // Validates expression to right of operator
-      var rsearch = true;
-      var op_r = j;
+      let rSearch = true;
+      let op_r = j;
       if (op_r > len - 1) {
         return false;
       }
-      while(rsearch) {
+      while(rSearch) {
         if (!(/[\s]/.test(str[op_r]))) {
           if (!(/[a-zA-Z0-9.'"(\\-]/.test(str[op_r]))) {
             return false;
           }
-          rsearch = false;
+          rSearch = false;
         } else {
           op_r += 1;
         }
@@ -147,22 +147,22 @@ export function validate(str, vars) {
   }
 
   // Function validation
-  var lcount = 0;
-  var rcount = 0;
-  var substr = "";
-  var instring = false;
+  let lCount = 0;
+  let rCount = 0;
+  let substr = "";
+  let inString = false;
   // Loop to checks parenthesis, identify non-operator strings and check strings
-  for (var i = 0; i < len; i++) {
-    instring = isinstring(str, i, instring);
+  for (let i = 0; i < len; i++) {
+    inString = isInString(str, i, inString);
 
-    if (!instring) {
-      if (str[i] == "(") {
-        lcount += 1;
+    if (!inString) {
+      if (str[i] === "(") {
+        lCount += 1;
       }
-      if (str[i] == ")") {
-        rcount += 1;
+      if (str[i] === ")") {
+        rCount += 1;
       }
-      if (rcount > lcount) {
+      if (rCount > lCount) {
         return false;
       }
       // If usable char, add to substring
@@ -171,7 +171,7 @@ export function validate(str, vars) {
       }
       // Checks if substring is valid
       if ((str[i + 1] == null || !(/[[a-zA-Z0-9]/.test(str[i + 1])))) {
-        if ((funs.includes(substr) && str[i + 1] == "(") || substr == "") {
+        if ((functions.includes(substr) && str[i + 1] === "(") || substr === "") {
           substr = "";
         } else if (vars.includes(substr) || (ops.includes(substr) || !(isNaN(substr)))) {
           substr = "";
@@ -183,7 +183,7 @@ export function validate(str, vars) {
   }
 
   // Parentheses must be balanced and quotes should end
-  return (lcount == rcount) && !instring;
+  return (lCount === rCount) && !inString;
 }
 
 /**
@@ -191,76 +191,84 @@ export function validate(str, vars) {
  * @param {string} str - inputted normal syntax expression
  * @returns expression with converted functions
  */
-export function convert(str) {
-  var count = 0;
-  var stringminusquotes = getstringminusquotes(str);
-  var stringparts = getstringparts(str);
+function convert(str) {
+  let count = 0;
+  let stringMinusQuotes = getStringMinusQuotes(str);
+  let stringParts = getStringParts(str);
 
-  if (stringminusquotes.includes("^")) {
-    var i = indexofoutsidequotes(str, stringparts, "^");
-    var base = lfind(str, i);
-    var power = rfind(str, i);
+  if (stringMinusQuotes.includes("^")) {
+    let i = indexOutsideQuotes(str, stringParts, "^");
+    let base = lFind(str, i);
+    let power = rFind(str, i);
     str =
-      str.slice(0, i - base.length) +
-      base.trim() +
-      ".power(" +
-      power.trim() +
-      ")" +
-      str.slice(i + power.length + 1);
+        str.slice(0, i - base.length) +
+        base.trim() +
+        ".power(" +
+        power.trim() +
+        ")" +
+        str.slice(i + power.length + 1);
     count += 1;
   }
-  if (stringminusquotes.includes("**")) {
-    var i = indexofoutsidequotes(str, stringparts, "**");
-    var base = lfind(str, i);
-    var power = rfind(str, i+1);
+
+  if (stringMinusQuotes.includes("**")) {
+    let i = indexOutsideQuotes(str, stringParts, "**");
+    let base = lFind(str, i);
+    let power = rFind(str, i+1);
     str =
-      str.slice(0, i - base.length) +
-      base.trim() +
-      ".power(" +
-      power.trim() +
-      ")" +
-      str.slice(i + power.length + 2);
+        str.slice(0, i - base.length) +
+        base.trim() +
+        ".power(" +
+        power.trim() +
+        ")" +
+        str.slice(i + power.length + 2);
     count += 1;
   }
-  for (let f = 0; f < funs2.length; f++) {
-    if (stringminusquotes.includes(funs2[f])) {
-      if (str[indexofoutsidequotes(str, stringparts, funs2[f]) - 1] != ".") {
-        str = funcappend(str, funs2[f]);
+
+  for (let f = 0; f < noArgumentFunctions.length; f++) {
+    if (stringMinusQuotes.includes(noArgumentFunctions[f])) {
+      if (str[indexOutsideQuotes(str, stringParts, noArgumentFunctions[f]) - 1] !== ".") {
+        str = funcAppend(str, noArgumentFunctions[f]);
         count += 1;
       }
     }
   }
-  if (stringminusquotes.includes("LOG")) {
-    str = logappend(str, "LOG");
+
+  if (stringMinusQuotes.includes("LOG")) {
+    str = logAppend(str, "LOG");
     count += 1;
   }
-  if (stringminusquotes.includes("log")) {
-    if (str[indexofoutsidequotes(str, stringparts, "log") - 1] != ".") {
-      str = logappend(str, "log");
+
+  if (stringMinusQuotes.includes("log")) {
+    if (str[indexOutsideQuotes(str, stringParts, "log") - 1] !== ".") {
+      str = logAppend(str, "log");
       count += 1;
     }
   }
+
   if (str.includes("OR")) {
-    str = replaceoutsidequotes(str, stringparts, "OR", "or");
+    str = replaceOutsideQuotes(str, stringParts, "OR", "or");
     count += 1;
   }
+
   if (str.includes("AND")) {
-    str = replaceoutsidequotes(str, stringparts, "AND", "and");
+    str = replaceOutsideQuotes(str, stringParts, "AND", "and");
     count += 1;
   }
+
   if (str.includes("||")) {
-    str = replaceoutsidequotes(str, stringparts, "||", "or");
+    str = replaceOutsideQuotes(str, stringParts, "||", "or");
     count += 1;
   }
+
   if (str.includes("&&")) {
-    str = replaceoutsidequotes(str, stringparts, "&&", "and");
+    str = replaceOutsideQuotes(str, stringParts, "&&", "and");
     count += 1;
   }
 
   // Replace double quotes with single quotes
-  str = replaceoutsidequotes(str, stringparts, '"', "'", true);
+  str = replaceOutsideQuotes(str, stringParts, '"', "'", true);
 
-  if (count != 0) {
+  if (count !== 0) {
     return convert(str);
   } else {
     return str;
@@ -273,65 +281,65 @@ export function convert(str) {
  * @param {string} func - function in inputted normal syntax expression
  * @returns expression with converted function
  */
-export function funcappend(str, func) {
-  var i = str.indexOf(func);
-  var j = i + func.length;
-  var k = j;
-  var eq = false;
-  var open = 0;
-  var close = 0;
+function funcAppend(str, func) {
+  let i = str.indexOf(func);
+  let j = i + func.length;
+  let k = j;
+  let eq = false;
+  let open = 0;
+  let close = 0;
   while (!eq) {
-    if (str[k] == "(") {
+    if (str[k] === "(") {
       open += 1;
     }
-    if (str[k] == ")") {
+    if (str[k] === ")") {
       close += 1;
     }
 
-    if (open == close) {
+    if (open === close) {
       eq = true;
     } else {
       k += 1;
     }
   }
   return (
-    str.slice(0, i).trim() +
-    str.slice(j, k + 1).trim() +
-    "." +
-    func.toLowerCase() +
-    "()" +
-    str.slice(k + 1).trim()
+      str.slice(0, i).trim() +
+      str.slice(j, k + 1).trim() +
+      "." +
+      func.toLowerCase() +
+      "()" +
+      str.slice(k + 1).trim()
   );
 }
 
 /**
- * Same as funcappend, but in LOG format
+ * Same as funcAppend, but in LOG format
  * @param {string} str - inputted normal syntax expression
  * @param {string} func - "LOG" or "log"
  * @returns expression with converted log function
  */
-export function logappend(str, func) {
-  var stringparts = getstringparts(str);
-  var i = indexofoutsidequotes(str, stringparts, func);
-  var j = i + 3;
-  var k = j;
-  var cma = -1;
-  var eq = false;
-  var open = 0;
-  var close = 0;
+function logAppend(str, func) {
+  let stringParts = getStringParts(str);
+  let i = indexOutsideQuotes(str, stringParts, func);
+  let j = i + 3;
+  let k = j;
+  let cma = -1;
+  let eq = false;
+  let open = 0;
+  let close = 0;
 
   while (!eq) {
-    if (str[k] == "(") {
+    if (str[k] === "(") {
       open += 1;
     }
-    if (str[k] == ")") {
+    if (str[k] === ")") {
       close += 1;
     }
 
-    if (open == close + 1 && k != j && str[k] == ",") {
+    if (open === close + 1 && k !== j && str[k] === ",") {
       cma = k;
     }
-    if (open == close) {
+    if (open === close) {
       eq = true;
     } else {
       k += 1;
@@ -339,14 +347,14 @@ export function logappend(str, func) {
   }
 
   return (
-    str.slice(0, i).trim() +
-    "(" +
-    str.slice(cma + 1, k).trim() +
-    ")" +
-    ".log(" +
-    str.slice(j + 1, cma).trim() +
-    ")" +
-    str.slice(k + 1).trim()
+      str.slice(0, i).trim() +
+      "(" +
+      str.slice(cma + 1, k).trim() +
+      ")" +
+      ".log(" +
+      str.slice(j + 1, cma).trim() +
+      ")" +
+      str.slice(k + 1).trim()
   );
 }
 
@@ -356,36 +364,36 @@ export function logappend(str, func) {
  * @param {int} i - operator index
  * @returns expression to left of operator
  */
-export function lfind(str, i) {
-  if (str[i - 1] != ")") {
-    var search = true;
-    var lstr = "";
+function lFind(str, i) {
+  if (str[i - 1] !== ")") {
+    let search = true;
+    let lStr = "";
     while (search) {
       if (i < 2) {
         search = false;
       }
-      if (/[a-zA-Z0-9.-\s]/.test(str[i - 1])) {
-        lstr = str[i - 1] + lstr;
+      if (/[a-zA-Z0-9.\-\s]/.test(str[i - 1])) {
+        lStr = str[i - 1] + lStr;
         i -= 1;
       } else {
         search = false;
       }
     }
-    return lstr;
+    return lStr;
   } else {
-    var eq = false;
-    var open = 0;
-    var close = 0;
-    var k = i - 1;
+    let eq = false;
+    let open = 0;
+    let close = 0;
+    let k = i - 1;
 
     while (!eq) {
-      if (str[k] == "(") {
+      if (str[k] === "(") {
         open += 1;
       }
-      if (str[k] == ")") {
+      if (str[k] === ")") {
         close += 1;
       }
-      if (open == close) {
+      if (open === close) {
         eq = true;
       } else {
         k -= 1;
@@ -401,22 +409,22 @@ export function lfind(str, i) {
  * @param {int} i - operator index
  * @returns expression to right of operator
  */
-export function rfind(str, i) {
-  if (str[i + 1] != "(") {
-    var search = true;
-    var rstr = "";
+function rFind(str, i) {
+  if (str[i + 1] !== "(") {
+    let search = true;
+    let rStr = "";
     while (search) {
-      if (str[i + 2] == undefined) {
+      if (str[i + 2] === undefined) {
         search = false;
       }
       if (/[a-zA-Z0-9.\s()\\-]/.test(str[i + 1])) {
-        rstr = rstr + str[i + 1];
+        rStr = rStr + str[i + 1];
         i += 1;
       } else {
         search = false;
       }
     }
-    return rstr;
+    return rStr;
   } else {
     return str.slice(i + 1, str.slice(i).indexOf(")") + i + 1);
   }
@@ -428,11 +436,11 @@ export function rfind(str, i) {
  * @param {Array} vars - array of usable variables
  * @returns converted expression with formatted variables
  */
-export function varfind(str, vars) {
-  var end = false;
-  var i = 0;
-  var j = 0;
-  var v = "";
+function varFind(str, vars) {
+  let end = false;
+  let i = 0;
+  let j = 0;
+  let v = "";
   while (!end) {
     if (str[i] == null) {
       end = true;
@@ -465,21 +473,22 @@ export function varfind(str, vars) {
  * position
  * @param str {string} - string to check
  * @param i {number} - current string index
- * @param instring {boolean} - current status for in string
- * @return {boolean} - in string
+ * @param inString {boolean} - current status for in string
+ * @return {boolean} - True if the position i is inside a string
  */
-function isinstring(str, i, instring) {
-  var isquote = str[i] == "'" || str[i] =='"';
+function isInString(str, i, inString) {
+  const isQuote = str[i] === "'" || str[i] === '"';
 
-  if (isquote && !instring) {
-    // Check for quote start
+  if (isQuote && !inString) {  // Check for quote start
     return true;
-  } else if (isquote && instring &&
-      ((i == 0) || (i > 0 && str[i - 1] != '\\'))) {
+  } else if (isQuote && inString &&
+      ((i === 0) || (i > 0 && str[i - 1] !== "\\") ||
+       (i > 1 && str[i - 1] === "\\" && str[i - 2] === "\\"))) {
     // Check for quote end unless the quote is escaped
+    // Also check if the escape is escaped itself
     return false;
   } else {
-    return instring;
+    return inString;
   }
 }
 
@@ -489,62 +498,63 @@ function isinstring(str, i, instring) {
  * @return Array of booleans representing if position is part of quotes (which
  *  should not be processed)
  */
-function getstringparts(str) {
-  var parts = [];
-  var instring = false;
+function getStringParts(str) {
+  const parts = [];
+  let inString = false;
 
-  for (var i = 0; i < str.length; i++) {
-    instring = isinstring(str, i, instring);
+  for (let i = 0; i < str.length; i++) {
+    inString = isInString(str, i, inString);
 
-    parts.push(instring);
+    parts.push(inString);
   }
 
   return parts;
 }
 
 /**
- * Get only the parts of the string which do are not quotes
+ * Get only the parts of the string which are not quoted
  * @param str {string} - String to process
  * @return {string} - String without quotes
  */
-function getstringminusquotes(str) {
-  var parts = [];
-  var instring = false;
+function getStringMinusQuotes(str) {
+  const parts = [];
+  let inString = false;
 
-  for (var i = 0; i < str.length; i++) {
-    instring = isinstring(str, i, instring);
+  for (let i = 0; i < str.length; i++) {
+    inString = isInString(str, i, inString);
 
-    if (!instring) {
+    if (!inString) {
       parts.push(str[i]);
     }
   }
 
-  return parts.join('');
+  return parts.join("");
 }
 
 /**
  * Get the index for the search value but ignore quotes
  * @param str {string} - String to search
- * @param stringparts - Array of booleans indicating if inside string.
- *  Use `getstringparts`
+ * @param stringParts - Array of booleans indicating if inside string.
+ *  Use `getStringParts`
  * @param value {string} - Value to look for
  * @return {number} - Index of match, -1 if none (not including quotes)
  */
-function indexofoutsidequotes(str, stringparts, value) {
-  var insidequote = false;
-  var searchindex = 0;
+function indexOutsideQuotes(str, stringParts, value) {
+  let insideQuote = false;
+  let searchIndex = 0;
+  let index;
 
   do {
-    var index = str.indexOf(value, searchindex);
+    index = str.indexOf(value, searchIndex);
 
-    if (index !== -1 && stringparts[index]) {
+    if (index !== -1 && stringParts[index]) {
       // The index fell as part of a quote, we should find the next match
-      insidequote = true;
-      searchindex = index + 1;
+      insideQuote = true;
+      searchIndex = index + 1;
     } else {
-      insidequote = false;
+      insideQuote = false;
     }
-  } while (index !== -1 && insidequote);
+  } while (index !== -1 && insideQuote);
 
   return index;
 }
@@ -552,33 +562,34 @@ function indexofoutsidequotes(str, stringparts, value) {
 /**
  * Replace values not inside quotes
  * @param str {string} - String to search
- * @param stringparts - Array of booleans indicating if inside string.
- *  Use `getstringparts`
- * @param searchvalue {string} - Value to look for
- * @param replacevalue {string} - Value to replace with
- * @param replaceall {boolean} - Replace all matches. Default false.
+ * @param stringParts - Array of booleans indicating if inside string.
+ *  Use `getStringParts`
+ * @param searchValue {string} - Value to look for
+ * @param replaceValue {string} - Value to replace with
+ * @param replaceAll {boolean} - Replace all matches. Default false.
  * @return {string} - String with matches replaced outside of quotes
  */
-function replaceoutsidequotes(str, stringparts, searchvalue, replacevalue, replaceall = false) {
-  var insidequote = false;
-  var searchindex = 0;
-  var tmpstr = str.split('');
+function replaceOutsideQuotes(str, stringParts, searchValue, replaceValue, replaceAll = false) {
+  let insideQuote = false;
+  let searchIndex = 0;
+  let tmpStr = str.split("");
+  let index;
 
   do {
-    var index = str.indexOf(searchvalue, searchindex);
+    index = str.indexOf(searchValue, searchIndex);
 
-    if (index !== -1 && stringparts[index] && searchvalue !== '"') {
+    if (index !== -1 && stringParts[index] && searchValue !== '"') {
       // The index fell as part of a quote, we should find the next match
       // Special case when replacing surrounding quotes
-      insidequote = true;
-      searchindex = index + 1;
+      insideQuote = true;
+      searchIndex = index + 1;
     } else if (index !== -1) {
-      insidequote = false;
-      tmpstr.splice(index, searchvalue.length, ...replacevalue.split(''));
-      searchindex += searchvalue.length - replacevalue.length;
-      str = tmpstr.join('');
+      insideQuote = false;
+      tmpStr.splice(index, searchValue.length, ...replaceValue.split(""));
+      searchIndex += searchValue.length - replaceValue.length;
+      str = tmpStr.join("");
     }
-  } while (index !== -1 && replaceall);
+  } while (index !== -1 && replaceAll);
 
-  return tmpstr.join('');
+  return tmpStr.join("");
 }